fix a few C4267 x64 warnings
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@5307 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/muxer/DSMMuxer/DSMMuxer.cpp
@@ -353,7 +353,7 @@ void CDSMMuxerFilter::MuxPacket(IBitStream* pBS, const MuxerPacket* pPacket)
     pBS->BitWrite(iDuration, 3);
     pBS->BitWrite(myabs(rtTimeStamp), iTimeStamp << 3);
     pBS->BitWrite(rtDuration, iDuration << 3);
-    pBS->ByteWrite(pPacket->pData.GetData(), pPacket->pData.GetCount());
+    pBS->ByteWrite(pPacket->pData.GetData(), (int)pPacket->pData.GetCount());
 }
 
 void CDSMMuxerFilter::MuxFooter(IBitStream* pBS)
